requirements.txt
asgiref==3.9.1
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
kafka-python==2.2.15
packaging==25.0
sqlparse==0.5.3
psycopg[binary]==3.2.9
